+++ docs/CHANGELOG.md
@@ -1,8 +1,11 @@
-# v1 - Initial Release - 3/18/24
+# v1 - Initial Release
+
+3/18/24
 
 - Released the site.
 
-## v1.1 - First fixes 
+## v1.1 - First fixes
+
 3/19/24 - 3/21/24
 
 - Fixed game names and removed duplicates
@@ -14,11 +17,13 @@
 - Added easter egg
 
 ### v1.1.1 - SM64 Typo
+
 3/21/24
 
 - Renamed games/files/super-maio-64 to games/files/super-mario-64, which fixed a referencing error
 
 ## v1.2 - Search bar and game display page
+
 3/21/24
 
 - Made a page that displays the games and their title with option to go fullscreen or view the source code.
@@ -30,6 +35,7 @@
 - Made theme buttons look like their themes
 
 ## v1.3 - Categories
+
 3/22/24
 
 - Renamed minecraft games fom starting with mc to start with minecraft
@@ -38,11 +44,13 @@
 - Renamed games/game.html or games/display.html, renamed ?game query to ?g
 
 ### v1.3.1 - Anti-tab-close
+
 3/22/24
 
 - Made it so a teacher can't close the tab remotely.
 
 ## v1.4 - Alternate Links & Feedback
+
 3/23/24 - 3/25/24
 
 - Added an alternate links page
@@ -53,12 +61,14 @@
 - Made all game names lowercase
 
 ### v1.4.1 - Fixed game names
+
 3/25/24
 
 - Reordered game names in assets/js/games.js
 - Fixed some incorrect file names in development/CHANGELOG.md
 
-## v1.5 - Clock 
+## v1.5 - Clock
+
 3/25/24 - 4/3/24
 
 - Added new clock pages with interactive and animated backgrounds
@@ -73,31 +83,48 @@
 - Changed the layout of docs/CHANGELOG.md a little bit
 
 ### v1.5.1 - Incremancer
+
 4/4/24
 
 - Added incremancer to games
 
-#### v1.5.1.1 - Incremancer fix 
+#### v1.5.1.1 - Incremancer fix
+
 4/4/24
 
 - Tried to fix incremancer
 - Made navbar clock update faster
 
 #### v1.5.1.2 - Incremancer fix 2
+
 4/4/24
 
 - Tried to fix incremancer
 
 #### v1.5.1.3 - Rename Incremancer
+
 4/4/24
 
 - Renamed incremancer-master to incremancer
 
 ### v1.5.2 - Update Cookie Clicker
+
 4/9/24
 
 - Updated Cookie Clicker to newest version
 
-### v1.5.2.1 - Add note to try fullscreen
+#### v1.5.2.1 - Add note to try fullscreen
+
+4/9/24
 
 - Added note to try fullscreen
+
+## v1.6 - Unblocking
+
+4/13/24 - 4/17/24
+
+- Added about:blank embedder
+- Added random game
+- Added Monkey Mart
+- Added 667 new proxies
+- Added temperature and weather to clock pages